LayoutTests/animations/CSSKeyframesRule-name-null-expected.txt

Tests that the parameters are mandatory in CSSKeyframesRule API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS keyframes.__proto__ is CSSKeyframesRule.prototype
PASS keyframes.name is "test1"
PASS keyframes.name = null did not throw exception.
PASS keyframes.name is "null"
PASS successfullyParsed is true

TEST COMPLETE
